Split up user settings files
